+++ Harden-Windows-Security Module/Main files/C#/CimInstances/FirewallHelper.cs
