+++ Harden-Windows-Security Module/Main files/C#/CimInstances/GetEncryptedVolumeInfo.cs
