+++ Harden-Windows-Security Module/Main files/C#/CimInstances/MDM.cs
