+++ Harden-Windows-Security Module/Main files/C#/CimInstances/MpComputerStatusHelper.cs
