+++ Harden-Windows-Security Module/Main files/C#/CimInstances/MpPreferenceHelper.cs
