+++ Harden-Windows-Security Module/Main files/C#/CimInstances/NetConnectionProfiles.cs
