+++ Harden-Windows-Security Module/Main files/C#/CimInstances/TaskSchedulerHelper.cs
