+++ Harden-Windows-Security Module/Main files/C#/Types/CustomExceptions.cs
