+++ Harden-Windows-Security Module/Main files/C#/Types/DefenderPlatformUpdatesChannels.cs
