+++ Harden-Windows-Security Module/Main files/C#/Types/IndividualResultClass.cs
