+++ Harden-Windows-Security Module/Main files/C#/Windows APIs/FirmwareType.cs
