+++ Harden-Windows-Security Module/Main files/C#/Windows APIs/TPM.cs
